Start refactoring db accessor slightly to reduce boilerplate.
